LayoutTests/fast/url/urlpattern-invalid-pattern.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_throws_js(TypeError, () => { new URLPattern(new URL('https://example.org/%(')); } );
  assert_throws_js(TypeError, () => { new URLPattern(new URL('https://example.org/%((')); } );
}, `Test unclosed token`);
</script>
